fix bugs in blitpacket

+++ BlitPacket/Deciphering.cs
@@ -59,7 +59,7 @@ namespace BlitzBit {
         }
         public System.UInt16 GetUInt16 () { CheckBuffers();
 
-            System.UInt32 val = BitConverter.ToUInt16(byteBuffer, index); index += 2;
+            System.UInt16 val = BitConverter.ToUInt16(byteBuffer, index); index += 2;
 
             return val;
         }
